Fix windows path to bash
Git Bash does path expansion on /bin/bash which isn't being done with rusts Command, but since bash is already in the PATH, it will be found by invoking "bash". 
Alternative is to find what shell invoked the rust-parallel process, the parent process, but not sure how to get that.